+++ packages/client/src/devices/__tests__/SpeakerManager.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, beforeEach, describe, vi, it, expect } from 'vitest';
-import { mockAudioDevices } from './mocks';
+import { emitDeviceIds, mockAudioDevices, mockDeviceIds$ } from './mocks';
 import { of } from 'rxjs';
 import { SpeakerManager } from '../SpeakerManager';
 import { checkIfAudioOutputChangeSupported } from '../devices';
@@ -9,6 +9,7 @@ vi.mock('../devices.ts', () => {
   return {
     getAudioOutputDevices: vi.fn(() => of(mockAudioDevices)),
     checkIfAudioOutputChangeSupported: vi.fn(() => true),
+    deviceIds$: mockDeviceIds$(),
   };
 });
 
@@ -57,6 +58,16 @@ describe('SpeakerManager.test', () => {
     manager.setVolume(0.5);
 
     expect(manager.state.volume).toBe(0.5);
+  });
+
+  it('should disable device if selected device is disconnected', () => {
+    emitDeviceIds(mockAudioDevices);
+    const deviceId = mockAudioDevices[1].deviceId;
+    manager.select(deviceId);
+
+    emitDeviceIds(mockAudioDevices.slice(2));
+
+    expect(manager.state.selectedDevice).toBe('');
   });
 
   afterEach(() => {