+++ packages/react-sdk/src/core/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from './useDevices';
 export * from './useTrackElementVisibility';