+++ packages/react-sdk/src/types/components.ts
@@ -1,7 +1,0 @@
-import { ReactNode } from 'react';
-
-export type ChildrenOnly = { children: ReactNode };
-
-export type Readable<T> = {
-  [k in keyof T]: T[k];
-} & {};