+++ sample-apps/react/react-dogfood/components/Debug/useIsDebugMode.ts
@@ -1,0 +1,9 @@
+import { useSearchParams } from 'next/navigation';
+
+/**
+ * Internal purpose hook. Enables certain development mode tools.
+ */
+export const useIsDebugMode = () => {
+  const params = useSearchParams();
+  return !!params.get('debug');
+};