+++ sample-apps/react/react-video-demo/src/App.tsx
@@ -60,7 +60,8 @@ const Init = () => {
       tokenProvider,
       options: {
         logLevel:
-          log_level || import.meta.env.MODE === 'production' ? 'warn' : 'debug',
+          log_level ||
+          (import.meta.env.MODE === 'production' ? 'warn' : 'debug'),
       },
     });
     setClient(_client);
@@ -100,6 +101,9 @@ const Init = () => {
     if (!client) return;
     const call = client.call(callType, callId);
     setActiveCall(call);
+    call
+      .getOrCreate()
+      .catch((err) => console.error(`Failed to get or create call`, err));
 
     // @ts-ignore - for debugging
     window.call = call;