+++ packages/react-bindings/CHANGELOG.md
@@ -2,6 +2,31 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.3.8](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.3.7...@stream-io/video-react-bindings-0.3.8) (2023-11-16)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.4.8`
+### [0.3.7](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.3.6...@stream-io/video-react-bindings-0.3.7) (2023-11-13)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.4.7`
+
+### Features
+
+* **device-api:** Browser Permissions API ([#1184](https://github.com/GetStream/stream-video-js/issues/1184)) ([a0b3573](https://github.com/GetStream/stream-video-js/commit/a0b3573b630ff8450953cdf1102fe722aea83f6f))
+
+### [0.3.6](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.3.5...@stream-io/video-react-bindings-0.3.6) (2023-11-13)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.4.6`
+### [0.3.5](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.3.4...@stream-io/video-react-bindings-0.3.5) (2023-11-07)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.4.5`
 ### [0.3.4](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.3.3...@stream-io/video-react-bindings-0.3.4) (2023-11-02)
 
 ### Dependency Updates