+++ packages/styling/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.1.14](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.1.13...@stream-io/video-styling-0.1.14) (2023-11-07)
+
+
+### Bug Fixes
+
+* lift the debug helpers from the SDK to Pronto ([#1182](https://github.com/GetStream/stream-video-js/issues/1182)) ([8f31efc](https://github.com/GetStream/stream-video-js/commit/8f31efc71d9f85ef147d21b42f23876599c36072))
+
 ### [0.1.13](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.1.12...@stream-io/video-styling-0.1.13) (2023-10-18)